voterBalance was using a factor of the total tokens
voterBalance is now using a minimum balance (e.g:1)

netparamas *MinVoterBalancer is not an int

+++ governance/proposal.go
@@ -15,7 +15,7 @@ type ProposalParameters struct {
 	RequiredParticipation float64
 	RequiredMajority      float64
 	MinProposerBalance    uint64
-	MinVoterBalance       float64
+	MinVoterBalance       uint64
 }
 
 // ToEnact wraps the proposal in a type that has a convenient interface
